Stopping XNA Dev. Made v1.1 Win Forms version.
XNA Dev doesn't appear to be worth the trouble anymore. While it was
easy to get 80% of the functionality, the Fog Updates proved to require
System.Drawing.Image instances to properly draw a polygon. A lot of time
was wasted trying to figure out how to draw a polygon in the 3D world to
then save the Texture, but ultimately it's not something you can just
"quickly put together". All things considered, the WinForms solution is
sufficient for now, albeit may suffer from some problems with high
resolution images.