Frame Animation
Added a basic Frame Animation (gif basically) that is currently being
tested with a Megaman 3-frame standing still sprite, in combination with
the Translation Animation, to make the menu. Too bad the Megaman X
spritesheet is awful and too small... Attempts to zoom it just end up
blurry, as expected.