Moved element init outside of Connection logic
Because the control is hidden at that point, we need to get the
position/bounding box before we hide the element. Also forced 0,0
absolute positioning of the canvas, for logical reasons but also updated
the position-dependent code to be dynamic, so if we decide to move the
canvas later, it should just work. Updated demo/release zip as well.